python_client_v2/requirements.txt
# file: python_client/requirements.txt

web3
pycryptodome
pyshamir
requests
umbral
